requirements.txt
# InsightForge — pinned dependencies
# Use with: python -m venv .venv && source .venv/bin/activate && pip install -r requirements.txt

numpy>=1.26.0,<3
pandas>=2.2.0,<3
tabulate>=0.9.0,<1
python-dotenv>=1.0.0,<2
openai>=1.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
streamlit>=1.39.0
plotly>=5.18.0
kaleido>=0.2.0
